Preference changes in OS X plug-ins.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@87024 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/plugins/DFNet/macsrc/PluginConfigController.m
@@ -101,7 +101,7 @@ void ReadConfig()
 	[defaults registerDefaults:[NSDictionary dictionaryWithObjectsAndKeys:
 								[NSDictionary dictionaryWithObjectsAndKeys:
 								 @"127.0.0.1", kIPADDRKEY,
-								 @((unsigned short)33306), kIPPORT,
+								 @33306, kIPPORT,
 								 @1, kPLAYERNUM,
 								 nil], PrefsKey, nil]];
